% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-img-sigma.R
\name{img_sigma}
\alias{img_sigma}
\title{Persistence Surface parameters}
\usage{
img_sigma(range = c(unknown(), unknown()), trans = transform_log10())
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
Parameters related to vectorized persistence surfaces (images).
}
\details{
The persistence surface vectorization deploys
\code{\link[TDAvec:computePersistenceImage]{TDAvec::computePersistenceImage()}}.
See there for definitions and references.
}
\examples{
data.frame(dist = I(list(eurodist, UScitiesD))) |> 
  transform(pd = I(lapply(dist, ripserr::vietoris_rips))) |> 
  subset(select = c(pd)) |> 
  print() -> pd_data

(sig_man <- img_sigma(range = c(10, 50), trans = NULL))
grid_regular(sig_man)

(sig_hom <- img_sigma() |> get_pers_frac_range(x = pd_data, hom_degree = 0))
grid_regular(sig_hom)

(sig_hom <- img_sigma() |> get_pers_frac_range(x = pd_data, hom_degree = 1))
grid_regular(sig_hom)
}
