% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mnist}
\alias{mnist}
\alias{mnist_train}
\alias{mnist_test}
\title{MNIST handwritten digits}
\format{
Two data frames of 600 and 100 rows, respectively, and 2 variables:
\describe{
\item{\code{digit}}{list column of 28 × 28 numeric matrices}
\item{\code{label}}{integer digit}
}
}
\source{
\url{http://yann.lecun.com/exdb/mnist/}
}
\usage{
mnist_train
mnist_test
}
\description{
This is a 1\% sample from the MNIST handwritten digit dataset.
}
\keyword{datasets}
