% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-num-level.R
\name{num_level}
\alias{num_level}
\title{Persistence Landscape parameters}
\usage{
num_level(range = c(1L, unknown()), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
Parameters related to vectorized persistence landscapes.
}
\details{
The persistence surface vectorization deploys
\code{\link[TDAvec:computePersistenceLandscape]{TDAvec::computePersistenceLandscape()}}.
See there for definitions and references.
}
\examples{
data.frame(dist = I(list(eurodist, UScitiesD))) |> 
  transform(pd = I(lapply(dist, ripserr::vietoris_rips))) |> 
  subset(select = c(pd)) |> 
  print() -> pd_data

(lev_man <- num_level(range = c(1, 24)))
grid_regular(lev_man)

(lev_dat <- num_level() |> get_pairs_max(x = pd_data))
grid_regular(lev_dat)

(lev_hom <- num_level() |> get_pairs_max(x = pd_data, hom_degrees = seq(2L)))
grid_regular(lev_hom)
}
