% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-phom-point-cloud.R, R/step-vpd-ecc-mv.R,
%   R/tdarec.R
\name{required_pkgs.step_phom_point_cloud}
\alias{required_pkgs.step_phom_point_cloud}
\alias{required_pkgs.step_vpd_ecc}
\alias{required_pkgs.step_tdarec}
\title{S3 methods for tracking which additional packages are needed for steps.}
\usage{
\method{required_pkgs}{step_phom_point_cloud}(x, ...)

\method{required_pkgs}{step_vpd_ecc}(x, ...)

\method{required_pkgs}{step_tdarec}(x, ...)
}
\arguments{
\item{x}{A recipe step.}
}
\value{
A character vector.
}
\description{
Recipe-adjacent packages always list themselves as a required package so that
the steps can function properly within parallel processing schemes.
}
\keyword{internal}
