% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-vpd-ecc.R
\name{step_vpd_ecc}
\alias{step_vpd_ecc}
\title{Euler characteristic curve vectorization of persistent homology}
\usage{
step_vpd_ecc(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  max_hom_degree = Inf,
  xseq = NULL,
  xmin = NULL,
  xmax = NULL,
  xlen = NULL,
  xby = NULL,
  columns = NULL,
  keep_original_cols = TRUE,
  skip = FALSE,
  id = rand_id("vpd_ecc")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose variables
for this step. See \code{\link[recipes:selections]{selections()}} for more details.}

\item{role}{For model terms created by this step, what analysis role should
they be assigned? By default, the new columns created by this step from
the original variables will be used as \emph{predictors} in a model.}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{max_hom_degree, xseq, xmin, xmax, xlen, xby}{Parameters transformed and
passed to {TDAvec}.}

\item{columns}{A character string of the selected variable names. This field
is a placeholder and will be populated once \code{\link[recipes:prep]{prep()}} is used.}

\item{keep_original_cols}{A logical to keep the original variables in the
output. Defaults to \code{FALSE}.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[recipes:bake]{bake()}}? While all operations are baked
when \code{\link[recipes:prep]{prep()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}
}
\value{
An updated version of \code{recipe} with the new step added to the
sequence of any existing operations.
}
\description{
The function \code{step_vpd_ecc()} creates a \emph{specification} of a
recipe step that will convert a list-column of 3-column matrix
representations of persistence data to a list-column of 1-row matrices of
vectorizations.
}
\details{
Persistent homology is usually encoded as birth--death pairs (barcodes or
diagrams), but the space of persistence data sets does not satisfy convenient
statistical properties. Such applications as hypothesis testing and machine
learning benefit from transformations of persistence data, often to Hilbert
spaces (vector spaces with inner products and induced metrics).

The Euler characteristic curve vectorization deploys
\code{\link[TDAvec:computeECC]{TDAvec::computeECC()}}.

The \code{hom_degree()} and \code{max_hom_degree} arguments determine the degree, or
the highest degree (starting with \code{0}), of the features to be transformed.
\code{xseq} will specify an entire discretization grid, while \code{xmin}, \code{xmax}, and
one of \code{xlen} and \code{xby} will construct a grid using \code{\link[base:seq]{base::seq()}}.
}
\section{Tuning Parameters}{
This step has 1 tuning parameters:
\itemize{
\item \code{max_hom_degree}: Maximum Homological Degree (type: integer, default: Inf)
}
}

\examples{
# toy data set
klein_sampler <- function(n, prob = .5) {
  if (rbinom(1, 1, prob) == 0) {
    tdaunif::sample_klein_flat(n)
  } else {
    tdaunif::sample_klein_tube(n)
  }
}
phom_data <- data.frame(
  id = LETTERS[seq(6L)],
  phom = I(lapply(
    replicate(6L, klein_sampler(60), simplify = FALSE),
    \(d) as.data.frame(ripserr::vietoris_rips(d))
  )),
  part = rep(c("train", "test"), each = 3)
)
print(phom_data)
head(phom_data$phom[[1]])

phom_train <- filter(phom_data, part == "train")
phom_test <- filter(phom_data, part == "test")

# guess maximum filtration parameter
phom_train$phom |> 
  lapply(\(d) d$death) |> 
  unlist() |> max() |> 
  print() -> max_threshold
# # create grid
# phom_seq <- seq(0, round(max_threshold * 2.01, digits = 2), .01)

# build preprocessing recipe with custom settings
phom_train \%>\%
  recipe() \%>\%
  update_role(id, new_role = "id") \%>\%
  step_vpd_ecc(phom, max_hom_degree = 2, xmax = max_threshold, xby = .01) \%>\%
  print() -> phom_rec
# tidy the prepped recipe step
tidy(phom_rec, number = 1)

# prepare recipe with training data
phom_rec \%>\%
  prep(training = phom_train, strings_as_factors = FALSE) \%>\%
  print() -> phom_prep
print(phom_prep)
# tidy the prepped recipe step
tidy(phom_prep, number = 1)

# build preprocessing recipe with default settings
phom_train \%>\%
  recipe() \%>\%
  update_role(id, new_role = "id") \%>\%
  step_vpd_ecc(phom) \%>\%
  prep(training = phom_train, strings_as_factors = FALSE) ->
  phom_prep2
print(phom_prep2)

# preprocess training data
juice(phom_prep2)

# preprocess testing data
bake(phom_test, object = phom_prep2)

# two-step recipe from sample to vectorization
sample_data <- data.frame(
  id = LETTERS[seq(6L)],
  sample = I(c(replicate(6L, klein_sampler(60), simplify = FALSE))),
  part = rep(c("train", "test"), each = 3L)
)
sample_train <- filter(sample_data, part == "train")
sample_test <- filter(sample_data, part == "test")

# build preprocessing recipe with default settings
sample_train \%>\%
  recipe() \%>\%
  update_role(id, new_role = "id") \%>\%
  step_phom_point_cloud(sample, engine = "ripserr") \%>\%
  step_vpd_ecc(sample_phom, keep_original_cols = FALSE) \%>\%
  prep(training = sample_train, strings_as_factors = FALSE) ->
  sample_rec
print(sample_rec)

# preprocess training data
juice(sample_rec)

# preprocess testing data
bake(sample_test, object = sample_rec)
}
