% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdarec.R
\docType{package}
\name{tdarec}
\alias{tdarec}
\alias{tdarec-package}
\title{A \emph{recipes} and \emph{dials} Extension for Persistent Homology}
\description{
The \emph{tdarec} package extends \emph{recipes} and \emph{dials} by providing
pre-processing steps with tunable parameters for computing persistent
homology of suitable data and for vectorizing persistent homology.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/corybrunson/tdarec}
  \item Report bugs at \url{https://github.com/corybrunson/tdarec/issues}
}

}
\author{
\strong{Maintainer}: Jason Cory Brunson \email{cornelioid@gmail.com}

}
