% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunable.R,
%   R/zzz-step-vpd-algebraic-functions.R,
%   R/zzz-step-vpd-betti-curve.R, R/zzz-step-vpd-complex-polynomial.R,
%   R/zzz-step-vpd-descriptive-statistics.R,
%   R/zzz-step-vpd-euler-characteristic-curve.R,
%   R/zzz-step-vpd-normalized-life-curve.R, R/zzz-step-vpd-persistence-block.R,
%   R/zzz-step-vpd-persistence-image.R, R/zzz-step-vpd-persistence-landscape.R,
%   R/zzz-step-vpd-persistence-silhouette.R,
%   R/zzz-step-vpd-persistent-entropy-summary.R,
%   R/zzz-step-vpd-tent-template-functions.R,
%   R/zzz-step-vpd-tropical-coordinates.R
\name{tunable_tdavec}
\alias{tunable_tdavec}
\alias{tunable.step_vpd_algebraic_functions}
\alias{tunable.step_vpd_betti_curve}
\alias{tunable.step_vpd_complex_polynomial}
\alias{tunable.step_vpd_descriptive_statistics}
\alias{tunable.step_vpd_euler_characteristic_curve}
\alias{tunable.step_vpd_normalized_life_curve}
\alias{tunable.step_vpd_persistence_block}
\alias{tunable.step_vpd_persistence_image}
\alias{tunable.step_vpd_persistence_landscape}
\alias{tunable.step_vpd_persistence_silhouette}
\alias{tunable.step_vpd_persistent_entropy_summary}
\alias{tunable.step_vpd_tent_template_functions}
\alias{tunable.step_vpd_tropical_coordinates}
\title{\code{tunable()} methods for tdavec package}
\usage{
\method{tunable}{step_vpd_algebraic_functions}(x, ...)

\method{tunable}{step_vpd_betti_curve}(x, ...)

\method{tunable}{step_vpd_complex_polynomial}(x, ...)

\method{tunable}{step_vpd_descriptive_statistics}(x, ...)

\method{tunable}{step_vpd_euler_characteristic_curve}(x, ...)

\method{tunable}{step_vpd_normalized_life_curve}(x, ...)

\method{tunable}{step_vpd_persistence_block}(x, ...)

\method{tunable}{step_vpd_persistence_image}(x, ...)

\method{tunable}{step_vpd_persistence_landscape}(x, ...)

\method{tunable}{step_vpd_persistence_silhouette}(x, ...)

\method{tunable}{step_vpd_persistent_entropy_summary}(x, ...)

\method{tunable}{step_vpd_tent_template_functions}(x, ...)

\method{tunable}{step_vpd_tropical_coordinates}(x, ...)
}
\arguments{
\item{x}{A recipe step object}

\item{...}{Not used.}
}
\value{
A tibble object.
}
\description{
These functions define what parameters \emph{can} be tuned for
specific steps. They also define the recommended objects from
\link[dials:dials-package]{dials} that can be used to generate new parameter
values and other characteristics.
}
\keyword{internal}
