% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpd-finalizers.R
\name{vpd-finalizers}
\alias{vpd-finalizers}
\alias{get_persistence_range}
\alias{get_pairs_max}
\title{Finalizers for persistent homology vectorizations}
\usage{
get_persistence_range(object, x, hom_degree = NULL, ...)

get_pairs_max(object, x, hom_degrees = NULL, ...)
}
\value{
An updated \code{param} object or a list of updated \code{param} objects depending
on what is provided in \code{object}.
}
\description{
These functions take a persistent homology vectorization
parameter object and modify the \code{\link[dials:unknown]{dials::unknown()}} parts of \code{ranges} based
on a data set and heuristics used in inaugural studies.
}
