% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-phom-point-cloud.R, R/tdarec.R,
%   R/zzz-step-vpd-ecc.R, R/zzz-step-vpd-nl.R, R/zzz-step-vpd-pes.R,
%   R/zzz-step-vpd-pi.R, R/zzz-step-vpd-pl.R, R/zzz-step-vpd-ps.R,
%   R/zzz-step-vpd-vab.R, R/zzz-step-vpd-vpb.R
\name{required_pkgs.step_phom_point_cloud}
\alias{required_pkgs.step_phom_point_cloud}
\alias{required_pkgs.step_tdarec}
\alias{required_pkgs.step_vpd_ecc}
\alias{required_pkgs.step_vpd_nl}
\alias{required_pkgs.step_vpd_pes}
\alias{required_pkgs.step_vpd_pi}
\alias{required_pkgs.step_vpd_pl}
\alias{required_pkgs.step_vpd_ps}
\alias{required_pkgs.step_vpd_vab}
\alias{required_pkgs.step_vpd_vpb}
\title{S3 methods for tracking which additional packages are needed for steps.}
\usage{
\method{required_pkgs}{step_phom_point_cloud}(x, ...)

\method{required_pkgs}{step_tdarec}(x, ...)

\method{required_pkgs}{step_vpd_ecc}(x, ...)

\method{required_pkgs}{step_vpd_nl}(x, ...)

\method{required_pkgs}{step_vpd_pes}(x, ...)

\method{required_pkgs}{step_vpd_pi}(x, ...)

\method{required_pkgs}{step_vpd_pl}(x, ...)

\method{required_pkgs}{step_vpd_ps}(x, ...)

\method{required_pkgs}{step_vpd_vab}(x, ...)

\method{required_pkgs}{step_vpd_vpb}(x, ...)
}
\arguments{
\item{x}{A recipe step.}
}
\value{
A character vector.
}
\description{
Recipe-adjacent packages always list themselves as a required package so that
the steps can function properly within parallel processing schemes.
}
\keyword{internal}
