% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-blur.R, R/step-phom-degree.R,
%   R/step-phom-point-cloud.R, R/step-phom-raster.R,
%   R/tdarec.R, R/zzz-step-vpd-algebraic-functions.R,
%   R/zzz-step-vpd-betti-curve.R, R/zzz-step-vpd-complex-polynomial.R,
%   R/zzz-step-vpd-descriptive-statistics.R,
%   R/zzz-step-vpd-euler-characteristic-curve.R,
%   R/zzz-step-vpd-normalized-life-curve.R, R/zzz-step-vpd-persistence-block.R,
%   R/zzz-step-vpd-persistence-image.R, R/zzz-step-vpd-persistence-landscape.R,
%   R/zzz-step-vpd-persistence-silhouette.R,
%   R/zzz-step-vpd-persistent-entropy-summary.R,
%   R/zzz-step-vpd-tent-template-functions.R,
%   R/zzz-step-vpd-tropical-coordinates.R
\name{required_pkgs.step_blur}
\alias{required_pkgs.step_blur}
\alias{required_pkgs.step_phom_degree}
\alias{required_pkgs.step_phom_point_cloud}
\alias{required_pkgs.step_phom_raster}
\alias{required_pkgs.step_tdarec}
\alias{required_pkgs.step_vpd_algebraic_functions}
\alias{required_pkgs.step_vpd_betti_curve}
\alias{required_pkgs.step_vpd_complex_polynomial}
\alias{required_pkgs.step_vpd_descriptive_statistics}
\alias{required_pkgs.step_vpd_euler_characteristic_curve}
\alias{required_pkgs.step_vpd_normalized_life_curve}
\alias{required_pkgs.step_vpd_persistence_block}
\alias{required_pkgs.step_vpd_persistence_image}
\alias{required_pkgs.step_vpd_persistence_landscape}
\alias{required_pkgs.step_vpd_persistence_silhouette}
\alias{required_pkgs.step_vpd_persistent_entropy_summary}
\alias{required_pkgs.step_vpd_tent_template_functions}
\alias{required_pkgs.step_vpd_tropical_coordinates}
\title{S3 methods for tracking which additional packages are needed for
steps.}
\usage{
\method{required_pkgs}{step_blur}(x, ...)

\method{required_pkgs}{step_phom_degree}(x, ...)

\method{required_pkgs}{step_phom_point_cloud}(x, ...)

\method{required_pkgs}{step_phom_raster}(x, ...)

\method{required_pkgs}{step_tdarec}(x, ...)

\method{required_pkgs}{step_vpd_algebraic_functions}(x, ...)

\method{required_pkgs}{step_vpd_betti_curve}(x, ...)

\method{required_pkgs}{step_vpd_complex_polynomial}(x, ...)

\method{required_pkgs}{step_vpd_descriptive_statistics}(x, ...)

\method{required_pkgs}{step_vpd_euler_characteristic_curve}(x, ...)

\method{required_pkgs}{step_vpd_normalized_life_curve}(x, ...)

\method{required_pkgs}{step_vpd_persistence_block}(x, ...)

\method{required_pkgs}{step_vpd_persistence_image}(x, ...)

\method{required_pkgs}{step_vpd_persistence_landscape}(x, ...)

\method{required_pkgs}{step_vpd_persistence_silhouette}(x, ...)

\method{required_pkgs}{step_vpd_persistent_entropy_summary}(x, ...)

\method{required_pkgs}{step_vpd_tent_template_functions}(x, ...)

\method{required_pkgs}{step_vpd_tropical_coordinates}(x, ...)
}
\arguments{
\item{x}{A recipe step.}
}
\value{
A character vector.
}
\description{
Recipe-adjacent packages always list themselves as a required
package so that the steps can function properly within parallel processing
schemes.
}
\keyword{internal}
