% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-PHom.R, R/step-VPD-ECC.R, R/tdarec.R
\name{required_pkgs.step_phom}
\alias{required_pkgs.step_phom}
\alias{required_pkgs.step_vpd_ecc}
\alias{required_pkgs.step_tdarec}
\title{S3 methods for tracking which additional packages are needed for steps.}
\usage{
\method{required_pkgs}{step_phom}(x, ...)

\method{required_pkgs}{step_vpd_ecc}(x, ...)

\method{required_pkgs}{step_tdarec}(x, ...)
}
\arguments{
\item{x}{A recipe step.}
}
\value{
A character vector.
}
\description{
Recipe-adjacent packages always list themselves as a required package so that
the steps can function properly within parallel processing schemes.
}
\keyword{internal}
