% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpd-finalizers.R
\name{vpd-finalizers}
\alias{vpd-finalizers}
\alias{get_pers_frac_range}
\alias{get_pairs_max}
\title{Finalizers for persistent homology vectorizations}
\usage{
get_pers_frac_range(
  object,
  x,
  hom_degree = NULL,
  log_vals = TRUE,
  frac = c(1/100, 1/10),
  ...
)

get_pairs_max(object, x, hom_degrees = NULL, ...)
}
\arguments{
\item{object}{A \code{param} object or a list of \code{param} objects.}

\item{x}{Persistence data in a recognizable format.}

\item{hom_degree, hom_degrees}{Integer (vector) of homological degree(s).}

\item{log_vals}{A logical: should the ranges be set on the log10 scale?}

\item{frac}{A double for the fraction of the data to be used for the upper
bound. For \code{get_n_frac_range()} and \code{get_batch_sizes()}, a vector of two
fractional values are required.}

\item{...}{Other arguments to pass to the underlying parameter finalizer
functions.}
}
\value{
An updated \code{param} object or a list of updated \code{param} objects depending
on what is provided in \code{object}.
}
\description{
These functions take a persistent homology vectorization
parameter object and modify the \code{\link[dials:unknown]{dials::unknown()}} parts of \code{ranges} based
on a data set and heuristics used in inaugural studies.
}
