% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpd-summarizers.R
\name{vpd-summarizers}
\alias{vpd-summarizers}
\alias{ph_dim}
\alias{ph_dim.default}
\alias{ph_dim.matrix}
\alias{ph_dim.array}
\alias{ph_dim.data.frame}
\alias{ph_dim.dist}
\alias{ph_dim.ts}
\alias{pairs_max}
\alias{pairs_max.default}
\alias{pairs_max.matrix}
\alias{pairs_max.data.frame}
\alias{pairs_max.diagram}
\alias{pairs_max.PHom}
\alias{pairs_max.persistence}
\alias{persistence_range}
\alias{persistence_range.default}
\alias{persistence_range.matrix}
\alias{persistence_range.data.frame}
\alias{persistence_range.diagram}
\alias{persistence_range.PHom}
\alias{persistence_range.persistence}
\alias{union_range}
\alias{union_range.default}
\alias{union_range.matrix}
\alias{union_range.data.frame}
\alias{union_range.diagram}
\alias{union_range.PHom}
\alias{union_range.persistence}
\title{Summarize topological data}
\usage{
ph_dim(x)

\method{ph_dim}{default}(x)

\method{ph_dim}{matrix}(x)

\method{ph_dim}{array}(x)

\method{ph_dim}{data.frame}(x)

\method{ph_dim}{dist}(x)

\method{ph_dim}{ts}(x)

pairs_max(x, hom_degrees)

\method{pairs_max}{default}(x, hom_degrees)

\method{pairs_max}{matrix}(x, hom_degrees)

\method{pairs_max}{data.frame}(x, hom_degrees)

\method{pairs_max}{diagram}(x, hom_degrees)

\method{pairs_max}{PHom}(x, hom_degrees)

\method{pairs_max}{persistence}(x, hom_degrees)

persistence_range(x, hom_degree)

\method{persistence_range}{default}(x, hom_degree)

\method{persistence_range}{matrix}(x, hom_degree)

\method{persistence_range}{data.frame}(x, hom_degree)

\method{persistence_range}{diagram}(x, hom_degree)

\method{persistence_range}{PHom}(x, hom_degree)

\method{persistence_range}{persistence}(x, hom_degree)

union_range(x, hom_degree)

\method{union_range}{default}(x, hom_degree)

\method{union_range}{matrix}(x, hom_degree)

\method{union_range}{data.frame}(x, hom_degree)

\method{union_range}{diagram}(x, hom_degree)

\method{union_range}{PHom}(x, hom_degree)

\method{union_range}{persistence}(x, hom_degree)
}
\description{
These miscellaneous functions are used by various
\verb{get_*_range()} functions to \link[dials:finalize]{finalize} hyperparameter
ranges.
}
\details{
The functions compute the following summaries:

\itemize{
\item \code{ph_dim()}:
Dimension of a data set for purposes of PH
\item \code{pairs_max()}:
Maximum number of persistent pairs of any degree
\item \code{persistence_range()}:
Range of positive finite persistence for a given degree
\item \code{union_range()}:
Union of birth--death ranges for a given degree
}
}
